# specs/003-schema-redesign/contracts/api.yaml
openapi: 3.0.3
info:
  title: Jump Ticketing API — Schema Redesign
  description: |
    REST API contracts for the schema redesign feature (003).
    Covers new endpoints for Organizations, Venues, Price Tiers, Orders,
    Users, and Guest Lookup. Updates existing Event and Ticket endpoints.
    Authentication moves to Auth.js v5 (JWT in Authorization header).
  version: 2.0.0
  contact:
    name: Jump Platform
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.jump.example.com/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /organizations/{organizationId}/venues:
    get:
      tags: [Venues]
      summary: List venues for an organization
      operationId: listVenues
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Venues retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  venues:
                    type: array
                    items:
                      $ref: "#/components/schemas/Venue"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Venues]
      summary: Create a venue
      operationId: createVenue
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueInput"
      responses:
        "201":
          description: Venue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/venues/{venueId}:
    get:
      tags: [Venues]
      summary: Get venue details
      operationId: getVenue
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/VenueIdParam"
      responses:
        "200":
          description: Venue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Venues]
      summary: Update a venue
      operationId: updateVenue
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/VenueIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueInput"
      responses:
        "200":
          description: Venue updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Venues]
      summary: Delete a venue
      description: Fails if the venue is referenced by existing events (FR-010).
      operationId: deleteVenue
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/VenueIdParam"
      responses:
        "204":
          description: Venue deleted
        "409":
          description: Venue is in use by events and cannot be deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===== EVENTS (FR-050, updated) =====
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth.js v5 JWT (HS256). Payload includes sub (userId), email, role.
  parameters:
    OrganizationIdParam:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
    VenueIdParam:
      name: venueId
      in: path
      required: true
      schema:
        type: string
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient role or organization access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: "Tier inventory exceeds event capacity"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # ─── Organization ───
...
    Venue:
      type: object
      properties:
        id:
          type: string
        organizationId:
          type: string
        name:
          type: string
        address:
          type: string
        timezone:
          type: string
          example: America/New_York
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateVenueInput:
      type: object
      required: [name, address]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        address:
          type: string
          minLength: 1
        timezone:
          type: string
          default: America/New_York
        isPublic:
          type: boolean
          default: true
    UpdateVenueInput:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        timezone:
          type: string
        isPublic:
          type: boolean

    # ─── Event ───
...